local storage saving tasks
Got this up and running.  I am still not sure how to solve the issue of things getting on the page yet.  Also not sure how to prevent the issue where multiple tasks of the same type can get created.  I believe push is causing the issue and may have to look for an alternative tomorrow.